requirements.txt
scipy
tqdm
pyyaml
pyquaternion
tensorboardX
scikit-learn
open3d